feat: auto test
