+++ __tests__/__fixtures__/x-if/actual.js
@@ -1,0 +1,9 @@
+import { createElement } from 'react';
+
+function Foo(props) {
+  return (
+    <View {...props} x-if={true} className="container">
+      <View x-if={condition}>First</View>
+    </View>
+  )
+}
