+++ jest.config.js
@@ -1,0 +1,8 @@
+// Sync object
+/** @type {import('@jest/types').Config.InitialOptions} */
+const config = {
+  verbose: true,
+  testMatch: ['!**/__fixtures__/**', '**/__tests__/**/*.js']
+};
+
+module.exports = config;
