+++ package.json
@@ -19,7 +19,7 @@
     "typescript": "^4.7.3"
   },
   "dependencies": {},
-  "author": "andycall",
+  "author": "jsx-plus <i@zeroling.com>",
   "license": "MIT",
   "bugs": {
     "url": "https://github.com/jsx-plus/swc-plugin-transform-jsx-condition/issues"
